Mark package as a conflict with zend-servicemanager v2
As noted in zendframework/zend-servicemanager#125, because this package
duplicates several classes from zend-servicemanager v2, it is technically a
conflict. This is noted in particular if you try do do `composer dumpautoload
--optimize`, as that command will fail due to failure to disambiguate.

As such, this patch removes the v2 compatibility within its requirements, and
explicitly marks v2 as a conflict.

Additionally, as this will be a breaking change, it also updates the package to
specify PHP 5.6 as the new minimum supported PHP version.